+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### features
 - fixes the double header issue (#17)
 - adds support for the ```gzip``` feature, which decompress the response content based on the content-type header
+- proper handling of graphql errors with a new `ParsedGraphqlResponse` type
 
 ### 0.7.1
 #### breaking change